[Refactoring] Extracted job scheduling logic from schedulers into a separate file
