+++ core/src/autogluon/core/scheduler/scheduler.py
@@ -1,42 +1,21 @@
 """Distributed Task Scheduler"""
-import contextlib
-import os
-import pickle
 import logging
-import shutil
+import pickle
 import sys
-import tempfile
-from functools import partial
-
-import dill
-import distributed
-from warnings import warn
-import multiprocessing as mp
 from collections import OrderedDict
+from warnings import warn
 
+import distributed
+
+from .jobs import DistributedJobRunner
 from .managers import TaskManagers
-from .remote import RemoteManager
-from .resource import DistributedResourceManager
 from .. import Task
-from .reporter import *
-from ..utils import AutoGluonWarning, AutoGluonEarlyStop, CustomProcess
-from ..utils.multiprocessing_utils import is_fork_enabled
-
-SYS_ERR_OUT_FILE = 'sys_err.out'
-SYS_STD_OUT_FILE = 'sys_std.out'
+from ..utils import AutoGluonWarning
 
 logger = logging.getLogger(__name__)
 
 __all__ = ['TaskScheduler']
 
-
-@contextlib.contextmanager
-def make_temp_directory():
-    temp_dir = tempfile.mkdtemp()
-    try:
-        yield temp_dir
-    finally:
-        shutil.rmtree(temp_dir)
 
 class ClassProperty(object):
 
@@ -51,6 +30,7 @@ class TaskScheduler(object):
     """Base Distributed Task Scheduler
     """
     managers = TaskManagers()
+    jobs = DistributedJobRunner()
 
     def __init__(self, dist_ip_addrs=None):
         cls = TaskScheduler
@@ -102,7 +82,7 @@ class TaskScheduler(object):
         cls = TaskScheduler
         if not task.resources.is_ready:
             cls.managers.request_resources(task.resources)
-        job = cls._start_distributed_job(task, cls.managers.resource_manager)
+        job = cls.jobs.start_distributed_job(task, cls.managers.resource_manager)
         new_dict = self._dict_from_task(task)
         new_dict['Job'] = job
         with self.managers.lock:
@@ -113,109 +93,8 @@ class TaskScheduler(object):
         """
         cls = TaskScheduler
         cls.managers.request_resources(task.resources)
-        job = cls._start_distributed_job(task, cls.managers.resource_manager)
+        job = cls.jobs.start_distributed_job(task, cls.managers.resource_manager)
         return job.result()
-
-    @staticmethod
-    def _start_distributed_job(task, resource_manager):
-        """Async Execute the job in remote and release the resources
-        """
-        logger.debug('\nScheduling {}'.format(task))
-        job = task.resources.node.submit(TaskScheduler._run_dist_job,
-                                         task.fn, task.args, task.resources.gpu_ids)
-        def _release_resource_callback(fut):
-            logger.debug('Start Releasing Resource')
-            resource_manager._release(task.resources)
-        job.add_done_callback(_release_resource_callback)
-        return job
-
-
-    @staticmethod
-    def _wrapper(tempdir, task, *args, **kwargs):
-        with open(os.path.join(tempdir, SYS_STD_OUT_FILE), 'w') as std_out:
-            with open(os.path.join(tempdir, SYS_ERR_OUT_FILE), 'w') as err_out:
-                sys.stdout = std_out
-                sys.stderr = err_out
-                task(*args, **kwargs)
-
-    @staticmethod
-    def _wrap(tempdir, task):
-        return partial(TaskScheduler._wrapper, tempdir, task)
-
-    @staticmethod
-    def _worker(pickled_fn, pickled_args, return_list, gpu_ids, args):
-        """Worker function in the client
-        """
-
-        # Only fork mode allows passing non-picklable objects
-        fn = pickled_fn if is_fork_enabled() else dill.loads(pickled_fn)
-        args = {**pickled_args, **args} if is_fork_enabled() else {**dill.loads(pickled_args), **args}
-
-        if len(gpu_ids) > 0:
-            # handle GPU devices
-            os.environ['CUDA_VISIBLE_DEVICES'] = ",".join(map(str, gpu_ids))
-            os.environ['MXNET_CUDNN_AUTOTUNE_DEFAULT'] = "0"
-        else:
-            os.environ['CUDA_VISIBLE_DEVICES'] = "-1"
-
-        # running
-        try:
-            ret = fn(**args)
-        except AutoGluonEarlyStop:
-            ret = None
-        return_list.append(ret)
-
-
-    @staticmethod
-    def _run_dist_job(fn, args, gpu_ids):
-        """Remote function Executing the task
-        """
-        if '_default_config' in args['args']:
-            args['args'].pop('_default_config')
-
-        if 'reporter' in args:
-            local_reporter = LocalStatusReporter()
-            dist_reporter = args['reporter']
-            args['reporter'] = local_reporter
-
-        manager = mp.Manager()
-        return_list = manager.list()
-
-        try:
-            # Starting local process
-            # Note: we have to use dill here because every argument passed to a child process over spawn or forkserver
-            # has to be pickled. fork mode does not require this because memory sharing, but it is unusable for CUDA
-            # applications (CUDA does not support fork) and multithreading issues (hanged threads).
-            # Usage of decorators makes standard pickling unusable (https://docs.python.org/3/library/pickle.html#what-can-be-pickled-and-unpickled)
-            # Dill enables sending of decorated classes. Please note if some classes are used in the training function,
-            # those classes are best be defined inside the function - this way those can be constructed 'on-the-other-side'
-            # after deserialization.
-            pickled_fn = fn if is_fork_enabled() else dill.dumps(fn)
-
-            # Reporter has to be separated since it's used for cross-process communication and has to be passed as-is
-            args_ = {k: v for (k, v) in args.items() if k not in ['reporter']}
-            pickled_args = args_ if is_fork_enabled() else dill.dumps(args_)
-
-            cross_process_args = {k: v for (k, v) in args.items() if k not in ['fn', 'args']}
-
-            with make_temp_directory() as tempdir:
-                p = CustomProcess(
-                    target=TaskScheduler._wrap(tempdir, partial(TaskScheduler._worker, pickled_fn, pickled_args)),
-                    args=(return_list, gpu_ids, cross_process_args)
-                )
-                p.start()
-                if 'reporter' in args:
-                    cp = Communicator.Create(p, local_reporter, dist_reporter)
-                p.join()
-                # Get processes outputs
-                with open(os.path.join(tempdir, SYS_STD_OUT_FILE)) as f:
-                    print(f.read(), file=sys.stdout, end = '')
-                with open(os.path.join(tempdir, SYS_ERR_OUT_FILE)) as f:
-                    print(f.read(), file=sys.stderr, end = '')
-        except Exception as e:
-            logger.error('Exception in worker process: {}'.format(e))
-        ret = return_list[0] if len(return_list) > 0 else None
-        return ret
 
     def _clean_task_internal(self, task_dict):
         pass
@@ -291,5 +170,5 @@ class TaskScheduler(object):
 
     def __repr__(self):
         reprstr = self.__class__.__name__ + '(\n' + \
-            str(self.managers.resource_manager) +')\n'
+                  str(self.managers.resource_manager) + ')\n'
         return reprstr
